+++ Robust.Shared/Physics/BroadphaseComponent.cs
@@ -1,4 +1,5 @@
 using Robust.Shared.GameObjects;
+using Robust.Shared.GameStates;
 using Robust.Shared.Physics.BroadPhase;
 
 namespace Robust.Shared.Physics
@@ -6,7 +7,7 @@ namespace Robust.Shared.Physics
     /// <summary>
     /// Stores the broadphase structure for the relevant grid / map.
     /// </summary>
-    [RegisterComponent]
+    [RegisterComponent, NetworkedComponent]
     public sealed partial class BroadphaseComponent : Component
     {
         /// <summary>